Use long double

+++ numerical-integration/include/integration/integration-methods.h
@@ -1,10 +1,10 @@
 #ifndef INTEGRATION_H
 #define INTEGRATION_H
 
-typedef double (*function)(double);
+typedef long double (*function)(long double);
 
-double simpsons_rule(function func, double a, double b, int n);
-double left_riemann_sum(function func, double a, double b, int n);
-double middle_riemann_sum(function func, double a, double b, int n);
+long double simpsons_rule(function func, long double a, long double b, int n);
+long double left_riemann_sum(function func, long double a, long double b, int n);
+long double middle_riemann_sum(function func, long double a, long double b, int n);
 
 #endif // INTEGRATION_H
