+++ numerical-integration/include/runges-rule/runges-rule.h
@@ -1,9 +1,9 @@
 #ifndef RUNGES_RULE_H
 #define RUNGES_RULE_H
 
-typedef double (*function)(double);
-typedef double (*integration_method)(function, double, double, int);
+typedef long double (*function)(long double);
+typedef long double (*integration_method)(function, long double, long double, int);
 
-double runges_rule(integration_method integration_meth, function func, double a, double b, int n);
+long double runges_rule(integration_method integration_meth, function func, long double a, long double b, int n);
 
 #endif //RUNGES_RULE_H
