+++ numerical-integration/src/integration/middle-riemann-sum.c
@@ -1,11 +1,11 @@
 #include "integration/integration-methods.h"
 
-double middle_riemann_sum(function func, double a, double b, int n) {
-    double h = (b - a) / n;
-    double sum = 0;
+long double middle_riemann_sum(function func, long double a, long double b, int n) {
+    long double h = (b - a) / n;
+    long double sum = 0;
 
     for (int i = 0; i < n; i++) {
-        double x = (a + i * h) + h / 2; // (x_i + x_(i+1)) / 2
+        long double x = (a + i * h) + h / 2; // (x_i + x_(i+1)) / 2
         sum += func(x);
     }
